backend/src/requirement.txt
Django==3.0.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
PyJWT==1.7.1